Move Azure DataGrid Provider into own class
