+++ extensions/azurecore/src/azureResource/azure-resource.d.ts
@@ -39,6 +39,7 @@ declare module 'azureResource' {
 			name: string;
 			id: string;
 			subscriptionId: string;
+			resourceGroup: string;
 			tenant?: string;
 		}
